requirements.txt
blinker==1.4
click==6.7
Flask==1.0.2
Flask-DebugToolbar==0.10.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.5
SQLAlchemy==1.2.8
google-cloud-translate==1.3.1
Flask_SocketIO
python-engineio
python-socketio
six==1.9.0
